+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to `erdemkeren/temporary-access` will be documented in this file.
 
+## 2.0.0 - 2017-01-04
+
+- The code and token discrimination removed from the package. Plain text and encrypted text used instead.
+- Totally separated token is now a part of access token.
+- Interface names were changed after 1.0.0 but never released. This version includes new interface names.
+- Most of the public service method signatures were changed due to removal of 'code'.
+
 ## 1.0.0 - 2016-12-30
 
 - First release of the package.